Add files via upload
This version is for developing standalone user-interface. Please note the standalone program in the zip file has some limitation because of the restrictions of MATLAB Compiler and MATLAB Coder (the external function/script cannot be executed through the standalone program).  I suggest users run the GMSSPro.m file (together with GMSSPro.fig file) to input self-defined attenuation function if required.