GMSSPro
The zip. package contains the GMSSPro standalone "exe" program (named GMSSProIstaller_Web) and the original GMSSPro. m. & fig. file which are used to generate the standalone program. The standalone program is not fully functioned because of the limitation of MATLAB Compiler. Users are suggested to use the GMSSPro.m and fig file within MATLAB to get the full functions. The "for_testing" file in the package can be used to test if your PC is eligible to install GMSSPro and can also be used for teaching or presentation purposes.