SSMainPro
SSMainPro.m file is designed for single-scenario ground motion simulation. The detailed illustration of each step is provided, and all the parameter value used in the program is displayed in the workspace in MATLAB. The author hopes this program can help users understand how stochastic ground motion simulation is implemented and how GMSS is programmed.