GMSSPro
GMSSPro is designed for user-friendly use. The MATLAB Runtime R2019b is required to be preinstalled in the computer. The file in "for testing" can be used for presentation or teaching purposes. Please note the standalone program (GMSSProInstaller_web.exe) is not fully functioned because of the limitation of MATLAB Compiler. Users are suggested to run the GMSSPro.m file (with GMSSPro.fig file) in MATLAB to get the full functions of GMSS.